Add common middleware mixin for controllers
