+++ app/controllers/middleware_deployment_controller.rb
@@ -1,19 +1,12 @@
 class MiddlewareDeploymentController < ApplicationController
   include EmsCommon
   include ContainersCommonMixin
+  include MiddlewareCommonMixin
 
   before_action :check_privileges
   before_action :get_session_data
   after_action :cleanup_action
   after_action :set_session_data
-
-  def show_list
-    process_show_list
-  end
-
-  def index
-    redirect_to :action => 'show_list'
-  end
 
   def show
     @display = params[:display] || "main" unless control_selected?
@@ -23,14 +16,4 @@ class MiddlewareDeploymentController < ApplicationController
     show_container(@record, controller_name, display_name)
   end
 
-  def listicon_image(item, _view)
-    icon = item.decorate.try(:listicon_image)
-    "100/#{icon}.png"
-  end
-
-  private ############################
-
-  def display_name
-    _("Middleware Deployments")
-  end
 end
