+++ app/controllers/middleware_server_controller.rb
@@ -1,6 +1,7 @@
 class MiddlewareServerController < ApplicationController
   include EmsCommon
   include ContainersCommonMixin
+  include MiddlewareCommonMixin
 
   before_action :check_privileges
   before_action :get_session_data
@@ -17,14 +18,6 @@ class MiddlewareServerController < ApplicationController
                                   :msg  => N_('Stop initiated for selected server(s)')
     }
   }.freeze
-
-  def show_list
-    process_show_list
-  end
-
-  def index
-    redirect_to :action => 'show_list'
-  end
 
   def show
     @display = params[:display] || "main" unless control_selected?
@@ -62,10 +55,6 @@ class MiddlewareServerController < ApplicationController
   end
 
   private ############################
-
-  def display_name
-    _('Middleware Servers')
-  end
 
   # Identify the selected servers. When we got the call from the
   # single server page, we need to look at :id, otherwise from
